fix: ChecklistItem.orderNo 타입을 SMALLINT에 맞춰 Short로 수정[#23]
checklist_items.order_no 컬럼이 SMALLINT인데 엔티티는 Integer로
매핑되어 있어 Hibernate 스키마 검증에 실패했다. Short로 수정하고
호출부/테스트도 함께 맞춘다.

+++ src/main/java/com/sayyoung/seed/domain/diagnosis/entity/ChecklistItem.java
@@ -35,7 +35,7 @@ public class ChecklistItem {
     private String contents;
 
     @Column(name = "order_no", nullable = false)
-    private Integer orderNo;
+    private Short orderNo;
 
     @Column(name = "estimated_amount", precision = 15, scale = 0)
     private BigDecimal estimatedAmount;
@@ -50,7 +50,7 @@ public class ChecklistItem {
             Recommendation recommendation,
             String itemKey,
             String contents,
-            Integer orderNo,
+            Short orderNo,
             BigDecimal estimatedAmount,
             String status
     ) {
@@ -69,7 +69,7 @@ public class ChecklistItem {
             Recommendation recommendation,
             String itemKey,
             String contents,
-            Integer orderNo,
+            Short orderNo,
             BigDecimal estimatedAmount
     ) {
         return new ChecklistItem(recommendation, itemKey, contents, orderNo, estimatedAmount, STATUS_TODO);
